pyproject.toml
[project]
name = "tb2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
]
